+++ R/intervals.r
@@ -64,14 +64,20 @@ setClass("Interval", contains = c("Timespan", "numeric"),
 
 #' @export
 setMethod("show", signature(object = "Interval"), function(object) {
-  print(format.Interval(object), quote = F)
+  if (length(object@.Data) == 0) {
+    cat("<Interval[0]>\n")
+  } else {
+    print(format(object), quote = FALSE)
+  }
 })
 
 #' @export
 format.Interval <- function(x, ...) {
-  if (length(x@.Data) == 0) return("Interval(0)")
-  paste(format(x@start, tz = x@tzone, usetz = TRUE), "--",
-    format(x@start + x@.Data, tz = x@tzone, usetz = TRUE), sep = "")
+  paste(
+    format(x@start, tz = x@tzone, usetz = TRUE),
+    format(x@start + x@.Data, tz = x@tzone, usetz = TRUE),
+    sep = "--"
+  )
 }
 
 #' @export