+++ tests/testthat/test-durations.R
@@ -4,6 +4,8 @@ test_that("duration() returns zero-length vector", {
   x <- duration()
   expect_s4_class(x, "Duration")
   expect_length(x, 0)
+  expect_equal(format(x), character())
+  expect_output(print(x), "<Duration[0]>", fixed = TRUE)
 })
 
 test_that("duration constructor doesn't accept non-numeric or non-character inputs", {
@@ -149,12 +151,6 @@ test_that("is.duration works as expected", {
   expect_true(is.duration(dminutes(1)))
   expect_false(is.duration(make_difftime(1000)))
   expect_false(is.duration(interval(lt_time, ct_time)))
-})
-
-test_that("format.Duration correctly displays intervals of length 0", {
-  dur <- duration(seconds = 5)
-
-  expect_output(print(dur[FALSE]), "Duration\\(0)")
 })
 
 test_that("format.Duration correctly displays durations with an NA", {